# Improve performance for didOpen method