--- a/src/clojure_lsp/feature/file_management.clj
+++ b/src/clojure_lsp/feature/file_management.clj
@@ -44,7 +44,7 @@
                          (assoc-in [:documents uri] {:v 0 :text text :saved-on-disk false})
                          (crawler/update-analysis uri (:analysis kondo-result))
                          (crawler/update-findings uri (:findings kondo-result)))))
-      (f.diagnostic/lint-file uri @db/db))))
+      (f.diagnostic/async-lint-file uri @db/db))))
 
 (defn ^:private find-changed-var-definitions [old-analysis new-analysis]
   (let [old-var-def (filter #(= :var-definitions (:bucket %)) old-analysis)
@@ -72,7 +72,7 @@
            (swap! db/db (fn [db] (-> db
                                      (crawler/update-analysis uri (:analysis new-analysis))
                                      (crawler/update-findings uri (:findings new-analysis)))))
-           (f.diagnostic/lint-file uri @db/db))))
+           (f.diagnostic/async-lint-file uri @db/db))))
      references-uri)))
 
 (defn ^:private offsets [lines line col end-line end-col]
@@ -129,7 +129,7 @@
                                                      (crawler/update-findings uri (:findings new-analysis))
                                                      (assoc :processing-changes false)))
               (do
-                (f.diagnostic/lint-file uri @db/db)
+                (f.diagnostic/sync-lint-file uri @db/db)
                 (when notify-references?
                   (notify-references old-analysis (get-in @db/db [:analysis filename]))))
               (recur @db/db))))))))