… docs/settings.md