@snoe 2021-05-25
good find, we absolutely should not be parsing the text ourselves ever - rely fully on kondo or cljfmt and push prs if need be.