@ericdallo 2021-05-25
@snoe we were parsing only for checking if the diagnostic was starting with a `(`, not sure how to delegate that to upstream